Cocoa: Fix isChildNSWindow() for embedded QWindows
Commit e3d21ef0 removed the manual m_isNSWindowChild
mode in favor of using m_view.window.parentWindow
as a declarative test.

However, this test has a false positive if the QWindow
in question is a non-child QWindow which is embedded
in a foreign NSView/NSWindow hierarchy. In the QTBUG-56589
case m_view.window.parentWindow is some internal window
in the Photos app, which is not relevant to the Qt
QWindow configuration.

Improve the isChildWindow() test: the QWindow must
actually be a child QWindow.

Task-number: QTBUG-56589
Change-Id: I0e6e262c85f47267a6165e7f98e48264b1e00069
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>