Cocoa: Add QNSView Metal Layer support
Make sure the layer is updated on window resize and
screen change.

Add a support test and print a warning if we try to
create a Metal layer without Metal system support.
This test should ideally be done earlier, before
configuring the QWindow to use Metal.

Link against the Metal framework: The minimum deployment
target is already 10.11 so this does not add additional
deployment requirements.

Change-Id: I0a38e824d0b6042bb52520dfaf0958ce21bb40b8
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>