QHeaderView: emit geometriesChanged() when the header is resized
When the QHeaderView is resized (e.g. due to setFixedWidth), the views
are not notified about it and the layout is not updated which leads to
a wrong painting. Therefore we have to listen for QEvent::Resize the
same way as when the font or style changes.

Task-number: QTBUG-67532
Task-number: QTBUG-34095
Change-Id: Id70daf083e0dddc2500d8e5aad47513266887a2c
Reviewed-by: Thorbjørn Lund Martsum <tmartsum@gmail.com>