Fix for input method commit that ends with newline
If the input method event contains a commit text that ends with a
newline, text, the commit string is inserted first. This changes the
current block.

This change makes sure that we apply the formatting changes (including
removing the old preedit text) to the old block in this specific case.

Task-number: QTBUG-67836
Change-Id: Ia83963780fb14b3c571dbbe3eb81fbbe20fbf412
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>